refactor(inspect): extract inspection levels, page readers, and output helpers
Split the inspect command into levels.go, pages.go, and helpers.go
with matching test files, keeping inspect.go focused on command
execution and the file summary.
