+++ cmd/inspect/inspect.go
@@ -2,21 +2,16 @@ package inspect
 
 import (
 	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"math"
-	"strings"
 
-	"github.com/hangxie/parquet-go/v3/common"
 	"github.com/hangxie/parquet-go/v3/parquet"
 	"github.com/hangxie/parquet-go/v3/reader"
-	"github.com/hangxie/parquet-go/v3/types"
 
 	pio "github.com/hangxie/parquet-tools/io"
 	pschema "github.com/hangxie/parquet-tools/schema"
 )
 
-// PageInfo represents information about a page in a Parquet file
+// PageInfo represents information about a page in a Parquet file.
 type PageInfo struct {
 	Index                      int               `json:"index"`
 	Offset                     int64             `json:"offset"`
@@ -149,440 +144,4 @@ func (c Cmd) inspectFile(reader *reader.ParquetReader) error {
 	}
 
 	return c.printJSON(output)
-}
-
-// Level 2: Row group details and column chunks with brief info
-func (c Cmd) inspectRowGroup(reader *reader.ParquetReader, rowGroupIndex int, inExNameMap map[string][]string, pathMap map[string]*pschema.SchemaNode, bloomSizeMap map[string]int32) error {
-	footer := reader.Footer
-
-	if rowGroupIndex < 0 || rowGroupIndex >= len(footer.RowGroups) {
-		return fmt.Errorf("row group index %d out of range [0, %d)", rowGroupIndex, len(footer.RowGroups))
-	}
-
-	rg := footer.RowGroups[rowGroupIndex]
-
-	// Build column chunks brief info
-	columnChunksBrief := make([]map[string]any, len(rg.Columns))
-	for i, col := range rg.Columns {
-		columnChunksBrief[i] = c.buildColumnChunkBrief(i, col, inExNameMap, pathMap, bloomSizeMap)
-	}
-
-	output := map[string]any{
-		"rowGroup": map[string]any{
-			"index":         rowGroupIndex,
-			"numRows":       rg.NumRows,
-			"totalByteSize": rg.TotalByteSize,
-			"numColumns":    len(rg.Columns),
-		},
-		"columnChunks": columnChunksBrief,
-	}
-
-	return c.printJSON(output)
-}
-
-func (c Cmd) buildColumnChunkBrief(index int, col *parquet.ColumnChunk, inExNameMap map[string][]string, pathMap map[string]*pschema.SchemaNode, bloomSizeMap map[string]int32) map[string]any {
-	pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-	pathInSchema := c.resolvePathInSchema(col.MetaData.PathInSchema, inExNameMap)
-	schemaNode := pathMap[pathKey]
-
-	columnChunk := map[string]any{
-		"index":            index,
-		"pathInSchema":     pathInSchema,
-		"type":             col.MetaData.Type.String(),
-		"encodings":        pschema.EncodingToString(col.MetaData.Encodings),
-		"compressionCodec": col.MetaData.Codec.String(),
-		"numValues":        col.MetaData.NumValues,
-		"compressedSize":   col.MetaData.TotalCompressedSize,
-		"uncompressedSize": col.MetaData.TotalUncompressedSize,
-	}
-
-	c.addTypeInformation(columnChunk, schemaNode)
-	c.addEncryptionInfo(columnChunk, col)
-
-	// Add bloom filter info if available, using correct bitset-only size
-	if col.MetaData.IsSetBloomFilterOffset() {
-		columnChunk["bloomFilterOffset"] = col.MetaData.GetBloomFilterOffset()
-		if size, ok := bloomSizeMap[pathKey]; ok && size > 0 {
-			columnChunk["bloomFilterLength"] = size
-		}
-	}
-
-	// Add statistics if available
-	if col.MetaData.Statistics != nil {
-		stats := c.buildStatistics(col.MetaData.Statistics, schemaNode)
-		if len(stats) > 0 {
-			columnChunk["statistics"] = stats
-		}
-	}
-
-	return columnChunk
-}
-
-// Level 3: Column chunk details and pages with brief info
-func (c Cmd) inspectColumnChunk(reader *reader.ParquetReader, rowGroupIndex, columnChunkIndex int, inExNameMap map[string][]string, pathMap map[string]*pschema.SchemaNode, bloomSizeMap map[string]int32) error {
-	footer := reader.Footer
-
-	if rowGroupIndex < 0 || rowGroupIndex >= len(footer.RowGroups) {
-		return fmt.Errorf("row group index %d out of range [0, %d)", rowGroupIndex, len(footer.RowGroups))
-	}
-
-	rg := footer.RowGroups[rowGroupIndex]
-
-	if columnChunkIndex < 0 || columnChunkIndex >= len(rg.Columns) {
-		return fmt.Errorf("column chunk index %d out of range [0, %d)", columnChunkIndex, len(rg.Columns))
-	}
-
-	col := rg.Columns[columnChunkIndex]
-	pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-	pathInSchema := c.resolvePathInSchema(col.MetaData.PathInSchema, inExNameMap)
-	schemaNode := pathMap[pathKey]
-
-	// Build column chunk details
-	columnChunkDetails := map[string]any{
-		"rowGroupIndex":    rowGroupIndex,
-		"columnChunkIndex": columnChunkIndex,
-		"pathInSchema":     pathInSchema,
-		"type":             col.MetaData.Type.String(),
-		"encodings":        pschema.EncodingToString(col.MetaData.Encodings),
-		"compressionCodec": col.MetaData.Codec.String(),
-		"numValues":        col.MetaData.NumValues,
-		"compressedSize":   col.MetaData.TotalCompressedSize,
-		"uncompressedSize": col.MetaData.TotalUncompressedSize,
-		"dataPageOffset":   col.MetaData.DataPageOffset,
-	}
-
-	if col.MetaData.DictionaryPageOffset != nil {
-		columnChunkDetails["dictionaryPageOffset"] = *col.MetaData.DictionaryPageOffset
-	}
-
-	if col.MetaData.IndexPageOffset != nil {
-		columnChunkDetails["indexPageOffset"] = *col.MetaData.IndexPageOffset
-	}
-
-	if col.MetaData.IsSetBloomFilterOffset() {
-		columnChunkDetails["bloomFilterOffset"] = col.MetaData.GetBloomFilterOffset()
-		if size, ok := bloomSizeMap[pathKey]; ok && size > 0 {
-			columnChunkDetails["bloomFilterLength"] = size
-		}
-	}
-
-	c.addTypeInformation(columnChunkDetails, schemaNode)
-	c.addEncryptionInfo(columnChunkDetails, col)
-
-	// Add statistics
-	if col.MetaData.Statistics != nil {
-		stats := c.buildStatistics(col.MetaData.Statistics, schemaNode)
-		if len(stats) > 0 {
-			columnChunkDetails["statistics"] = stats
-		}
-	}
-
-	// Read pages - this requires reading the actual data from the file
-	pages, err := c.readPages(reader, rowGroupIndex, columnChunkIndex, schemaNode)
-	if err != nil {
-		return fmt.Errorf("failed to read pages: %w", err)
-	}
-
-	output := map[string]any{
-		"columnChunk": columnChunkDetails,
-		"pages":       pages,
-	}
-
-	return c.printJSON(output)
-}
-
-// Level 4: Page details and values
-func (c Cmd) inspectPage(reader *reader.ParquetReader, rowGroupIndex, columnChunkIndex, pageIndex int, pathMap map[string]*pschema.SchemaNode) error {
-	footer := reader.Footer
-
-	if rowGroupIndex < 0 || rowGroupIndex >= len(footer.RowGroups) {
-		return fmt.Errorf("row group index %d out of range [0, %d)", rowGroupIndex, len(footer.RowGroups))
-	}
-	rg := footer.RowGroups[rowGroupIndex]
-
-	if columnChunkIndex < 0 || columnChunkIndex >= len(rg.Columns) {
-		return fmt.Errorf("column chunk index %d out of range [0, %d)", columnChunkIndex, len(rg.Columns))
-	}
-	col := rg.Columns[columnChunkIndex]
-	pathKey := strings.Join(col.MetaData.PathInSchema, common.ParGoPathDelimiter)
-	schemaNode := pathMap[pathKey]
-
-	// Read pages
-	pages, err := c.readPages(reader, rowGroupIndex, columnChunkIndex, schemaNode)
-	if err != nil {
-		return fmt.Errorf("failed to read pages: %w", err)
-	}
-
-	if pageIndex < 0 || pageIndex >= len(pages) {
-		return fmt.Errorf("page index %d out of range [0, %d)", pageIndex, len(pages))
-	}
-	page := pages[pageIndex]
-
-	// Read page values
-	values, err := c.readPageValues(reader, rowGroupIndex, columnChunkIndex, col, schemaNode, pages, pageIndex)
-	if err != nil {
-		return fmt.Errorf("failed to read page values: %w", err)
-	}
-
-	output := map[string]any{
-		"page":   page,
-		"values": values,
-	}
-
-	return c.printJSON(output)
-}
-
-func (c Cmd) readPages(pr *reader.ParquetReader, rowGroupIndex, columnChunkIndex int, schemaNode *pschema.SchemaNode) ([]PageInfo, error) {
-	pageHeaders, err := pr.GetAllPageHeaders(rowGroupIndex, columnChunkIndex)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read page headers: %w", err)
-	}
-
-	// Convert PageHeaderInfo to our output format
-	pages := make([]PageInfo, len(pageHeaders))
-	for i, headerInfo := range pageHeaders {
-		pages[i] = c.convertPageHeaderInfo(headerInfo, schemaNode)
-	}
-
-	return pages, nil
-}
-
-// convertPageHeaderInfo converts PageHeaderInfo from parquet-go to our JSON output format
-func (c Cmd) convertPageHeaderInfo(headerInfo reader.PageHeaderInfo, schemaNode *pschema.SchemaNode) PageInfo {
-	pageInfo := PageInfo{
-		Index:            headerInfo.Index,
-		Offset:           headerInfo.Offset,
-		Type:             headerInfo.PageType,
-		CompressedSize:   headerInfo.CompressedSize,
-		UncompressedSize: headerInfo.UncompressedSize,
-	}
-
-	if headerInfo.HasCRC {
-		pageInfo.HasCrc = true
-		pageInfo.Crc = fmt.Sprintf("%08x", uint32(headerInfo.CRC))
-	}
-
-	switch headerInfo.PageType {
-	case parquet.PageType_DATA_PAGE:
-		pageInfo.NumValues = &headerInfo.NumValues
-		pageInfo.Encoding = &headerInfo.Encoding
-		pageInfo.DefinitionLevelEncoding = &headerInfo.DefLevelEncoding
-		pageInfo.RepetitionLevelEncoding = &headerInfo.RepLevelEncoding
-
-		if headerInfo.HasStatistics {
-			pageInfo.Statistics = c.buildStatistics(headerInfo.Statistics, schemaNode)
-		}
-
-	case parquet.PageType_DATA_PAGE_V2:
-		pageInfo.NumValues = &headerInfo.NumValues
-		pageInfo.NumNulls = &headerInfo.NumNulls
-		pageInfo.NumRows = &headerInfo.NumRows
-		pageInfo.Encoding = &headerInfo.Encoding
-		pageInfo.DefinitionLevelsByteLength = &headerInfo.DefLevelBytes
-		pageInfo.RepetitionLevelsByteLength = &headerInfo.RepLevelBytes
-		pageInfo.IsCompressed = headerInfo.IsCompressed
-
-		if headerInfo.HasStatistics {
-			pageInfo.Statistics = c.buildStatistics(headerInfo.Statistics, schemaNode)
-		}
-
-	case parquet.PageType_DICTIONARY_PAGE:
-		pageInfo.NumValues = &headerInfo.NumValues
-		pageInfo.Encoding = &headerInfo.Encoding
-		pageInfo.IsSorted = headerInfo.IsSorted
-
-	case parquet.PageType_INDEX_PAGE:
-		pageInfo.Note = "Index page (column index)"
-	}
-
-	return pageInfo
-}
-
-func (c Cmd) readPageValues(pr *reader.ParquetReader, rowGroupIndex, columnChunkIndex int, col *parquet.ColumnChunk, schemaNode *pschema.SchemaNode, pages []PageInfo, pageIndex int) ([]any, error) {
-	meta := col.MetaData
-
-	if pageIndex < 0 || pageIndex >= len(pages) {
-		return nil, fmt.Errorf("page index %d out of range [0, %d)", pageIndex, len(pages))
-	}
-
-	pageInfo := pages[pageIndex]
-
-	// Handle different page types
-	switch pageInfo.Type {
-	case parquet.PageType_DICTIONARY_PAGE:
-		return c.readDictionaryPageValues(pr, col, schemaNode, pageInfo)
-	case parquet.PageType_DATA_PAGE, parquet.PageType_DATA_PAGE_V2:
-		// Continue to process data pages below
-	default:
-		// For other page types (INDEX_PAGE, etc.), return empty
-		return []any{}, nil
-	}
-
-	// Calculate total values before this row group and in this row group
-	var valuesBeforeRG int64
-	for i := range rowGroupIndex {
-		valuesBeforeRG += pr.Footer.RowGroups[i].NumRows
-	}
-
-	// Create a fresh column reader to read the entire column
-	freshReader, err := reader.NewParquetColumnReader(pr.PFile, reader.WithNP(4))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create fresh reader: %w", err)
-	}
-	defer func() { _ = freshReader.ReadStop() }()
-
-	// Calculate total number of rows in the file
-	totalRows := int64(0)
-	for _, rg := range pr.Footer.RowGroups {
-		totalRows += rg.NumRows
-	}
-
-	// Read ALL values from the entire file for this column
-	allValuesInFile, _, _, err := freshReader.ReadColumnByIndex(int64(columnChunkIndex), totalRows)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read column values: %w", err)
-	}
-
-	// Extract only the values for this row group
-	rgStartIdx := valuesBeforeRG
-	rgEndIdx := min(valuesBeforeRG+meta.NumValues, int64(len(allValuesInFile)))
-
-	allValues := allValuesInFile[rgStartIdx:rgEndIdx]
-
-	// Calculate the start index for this page
-	var startIdx int64
-	for i := range pageIndex {
-		if pages[i].Type != parquet.PageType_DATA_PAGE && pages[i].Type != parquet.PageType_DATA_PAGE_V2 {
-			continue
-		}
-		if pages[i].NumValues == nil {
-			continue
-		}
-		startIdx += int64(*pages[i].NumValues)
-	}
-
-	// Extract values for just this page
-	if pageInfo.NumValues == nil {
-		return nil, fmt.Errorf("unable to get numValues for page")
-	}
-
-	endIdx := min(startIdx+int64(*pageInfo.NumValues), int64(len(allValues)))
-
-	// Convert values to appropriate JSON types
-	pageValues := make([]any, endIdx-startIdx)
-	for i := startIdx; i < endIdx; i++ {
-		pageValues[i-startIdx] = allValues[i]
-	}
-
-	return c.convertValuesToJSON(pageValues, schemaNode), nil
-}
-
-func (c Cmd) readDictionaryPageValues(pr *reader.ParquetReader, col *parquet.ColumnChunk, schemaNode *pschema.SchemaNode, pageInfo PageInfo) ([]any, error) {
-	meta := col.MetaData
-
-	values, err := pr.ReadDictionaryPageValues(pageInfo.Offset, meta.Codec, meta.Type)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read dictionary page values: %w", err)
-	}
-
-	return c.convertValuesToJSON(values, schemaNode), nil
-}
-
-// Helper functions
-
-// printJSON marshals data to JSON and prints it
-func (c Cmd) printJSON(data any) error {
-	buf, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-	fmt.Println(string(buf))
-	return nil
-}
-
-// convertValuesToJSON converts raw parquet values to JSON-friendly types
-func (c Cmd) convertValuesToJSON(values []any, schemaNode *pschema.SchemaNode) []any {
-	result := make([]any, len(values))
-	for i, val := range values {
-		result[i] = types.ConvertToJSONType(val, &schemaNode.SchemaElement)
-	}
-	return result
-}
-
-// resolvePathInSchema resolves internal path to external path using the schema map
-func (c Cmd) resolvePathInSchema(pathInSchema []string, inExNameMap map[string][]string) []string {
-	pathKey := strings.Join(pathInSchema, common.ParGoPathDelimiter)
-	if exPath, found := inExNameMap[pathKey]; found {
-		return exPath
-	}
-	return pathInSchema
-}
-
-// addEncryptionInfo adds encryptionMode and keyMetadata to the output map when present
-func (c Cmd) addEncryptionInfo(output map[string]any, col *parquet.ColumnChunk) {
-	cm := col.GetCryptoMetadata()
-	if cm == nil {
-		return
-	}
-	switch {
-	case cm.ENCRYPTION_WITH_FOOTER_KEY != nil:
-		output["encryptionMode"] = "FOOTER_KEY"
-	case cm.ENCRYPTION_WITH_COLUMN_KEY != nil:
-		output["encryptionMode"] = "COLUMN_KEY"
-		if km := cm.ENCRYPTION_WITH_COLUMN_KEY.GetKeyMetadata(); len(km) > 0 {
-			output["keyMetadata"] = base64.StdEncoding.EncodeToString(km)
-		}
-	}
-}
-
-// addTypeInformation adds converted and logical type information to the output map
-func (c Cmd) addTypeInformation(output map[string]any, schemaNode *pschema.SchemaNode) {
-	if schemaNode == nil {
-		return
-	}
-	if ct := schemaNode.ConvertedTypeString(); ct != "" {
-		output["convertedType"] = ct
-	}
-	if lt := schemaNode.LogicalTypeString(); lt != "" {
-		output["logicalType"] = lt
-	}
-}
-
-// buildStatistics creates a statistics map from parquet statistics
-func (c Cmd) buildStatistics(statistics *parquet.Statistics, schemaNode *pschema.SchemaNode) map[string]any {
-	stats := map[string]any{}
-
-	if statistics.NullCount != nil {
-		stats["nullCount"] = *statistics.NullCount
-	}
-	if statistics.DistinctCount != nil {
-		stats["distinctCount"] = *statistics.DistinctCount
-	}
-
-	if schemaNode != nil {
-		min, max := schemaNode.DecodeStatistics(statistics)
-		if min != nil {
-			stats["minValue"] = normalizeNegativeZero(min)
-		}
-		if max != nil {
-			stats["maxValue"] = normalizeNegativeZero(max)
-		}
-	}
-
-	return stats
-}
-
-// normalizeNegativeZero converts IEEE 754 negative zero to positive zero for consistent JSON output.
-func normalizeNegativeZero(v any) any {
-	switch f := v.(type) {
-	case float32:
-		if math.Signbit(float64(f)) && f == 0 {
-			return float32(0)
-		}
-	case float64:
-		if math.Signbit(f) && f == 0 {
-			return float64(0)
-		}
-	}
-	return v
 }
